backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pandas<2.2.0
pycaret==3.3.2
google-genai>=1.0.0
pydantic>=2.9.0
python-dotenv>=1.0.0
